Move classes to SafeHandles
