+++ src/Common/src/Interop/Windows/advapi32/Interop.CryptEncrypt.cs
@@ -3,7 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Runtime.InteropServices;
-using System.Security.Cryptography;
+using Microsoft.Win32.SafeHandles;
 
 internal partial class Interop
 {
