+++ src/Common/src/Interop/Windows/advapi32/Interop.CryptSignHash.cs
@@ -4,7 +4,7 @@
 
 using System;
 using System.Runtime.InteropServices;
-using System.Security.Cryptography;
+using Microsoft.Win32.SafeHandles;
 
 internal partial class Interop
 {
