+++ src/System.Security.Cryptography.Csp/src/System.Security.Cryptography.Csp.csproj
@@ -54,7 +54,6 @@
     <Compile Include="System\Security\Cryptography\CapiHelper.DSA.Windows.cs" />
     <Compile Include="System\Security\Cryptography\CapiHelper.Windows.cs" />
     <Compile Include="System\Security\Cryptography\CapiHelper.SymmetricKey.cs" />
-    <Compile Include="System\Security\Cryptography\SafeCryptoHandles.cs" />
     <Compile Include="System\Security\Cryptography\Utils.cs" />
     <Compile Include="Internal\Cryptography\BasicSymmetricCipherCsp.cs" />
     <Compile Include="$(CommonPath)\Internal\Cryptography\BasicSymmetricCipher.cs">
@@ -126,9 +125,6 @@
     <Compile Include="$(CommonPath)\Interop\Windows\advapi32\Interop.CryptSetKeyParam.cs">
       <Link>Common\Interop\Windows\advapi32\Interop.CryptSetKeyParam.cs</Link>
     </Compile>
-    <Compile Include="$(CommonPath)\Interop\Windows\advapi32\Interop.CryptSetProvParam.cs">
-      <Link>Common\Interop\Windows\advapi32\Interop.CryptSetProvParam.cs</Link>
-    </Compile>
     <Compile Include="$(CommonPath)\Interop\Windows\advapi32\Interop.CryptSignHash.cs">
       <Link>Common\Interop\Windows\advapi32\Interop.CryptSignHash.cs</Link>
     </Compile>
@@ -143,6 +139,15 @@
     </Compile>
     <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeHandleCache.cs">
       <Link>Common\Microsoft\Win32\SafeHandles\SafeHandleCache.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeHashHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeHashHandle.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeKeyHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeKeyHandle.cs</Link>
+    </Compile>
+    <Compile Include="$(CommonPath)\Microsoft\Win32\SafeHandles\SafeProvHandle.cs">
+      <Link>Common\Microsoft\Win32\SafeHandles\SafeProvHandle.cs</Link>
     </Compile>
   </ItemGroup>
   <ItemGroup>
