+++ src/System.Security.Cryptography.Csp/src/System/Security/Cryptography/DSACryptoServiceProvider.Windows.cs
@@ -2,9 +2,10 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Internal.NativeCrypto;
 using System.Diagnostics;
 using System.IO;
+using Internal.NativeCrypto;
+using Microsoft.Win32.SafeHandles;
 
 namespace System.Security.Cryptography
 {
